+++ packages/pwa/app/components/recommended-products/index.jsx
@@ -6,9 +6,15 @@
  */
 import React, {useEffect, useRef, useState} from 'react'
 import PropTypes from 'prop-types'
+import {useIntl} from 'react-intl'
+import {Button} from '@chakra-ui/react'
 import ProductScroller from '../../components/product-scroller'
 import useEinstein from '../../commerce-api/hooks/useEinstein'
 import useIntersectionObserver from '../../hooks/use-intersection-observer'
+import useWishlist from '../../hooks/use-wishlist'
+import {useToast} from '../../hooks/use-toast'
+import useNavigation from '../../hooks/use-navigation'
+import {API_ERROR_MESSAGE} from '../../constants'
 
 /**
  * A component for fetching and rendering product recommendations from the Einstein API
@@ -24,6 +30,10 @@ const RecommendedProducts = ({zone, recommender, products, title, shouldFetch, .
         sendClickReco,
         sendViewReco
     } = useEinstein()
+    const wishlist = useWishlist()
+    const toast = useToast()
+    const navigate = useNavigation()
+    const {formatMessage} = useIntl()
 
     const ref = useRef()
     const isOnScreen = useIntersectionObserver(ref, {useOnce: true})
@@ -98,21 +108,86 @@ const RecommendedProducts = ({zone, recommender, products, title, shouldFetch, .
         return null
     }
 
+    // TODO: DRY the wishlist handlers when intl
+    // provider is available globally
+    const addItemToWishlist = async (product) => {
+        try {
+            await wishlist.createListItem({
+                id: product.productId,
+                quantity: 1
+            })
+            toast({
+                title: formatMessage(
+                    {
+                        defaultMessage:
+                            '{quantity} {quantity, plural, one {item} other {items}} added to wishlist'
+                    },
+                    {quantity: 1}
+                ),
+                status: 'success',
+                action: (
+                    // it would be better if we could use <Button as={Link}>
+                    // but unfortunately the Link component is not compatible
+                    // with Chakra Toast, since the ToastManager is rendered via portal
+                    // and the toast doesn't have access to intl provider, which is a
+                    // requirement of the Link component.
+                    <Button variant="link" onClick={() => navigate('/account/wishlist')}>
+                        View
+                    </Button>
+                )
+            })
+        } catch {
+            toast({
+                title: formatMessage(
+                    {defaultMessage: '{errorMessage}'},
+                    {errorMessage: API_ERROR_MESSAGE}
+                ),
+                status: 'error'
+            })
+        }
+    }
+    const removeItemFromWishlist = async (product) => {
+        try {
+            await wishlist.removeListItemByProductId(product.productId)
+            toast({
+                title: formatMessage({defaultMessage: 'Item removed from wishlist'}),
+                status: 'success',
+                id: product.productId
+            })
+        } catch {
+            toast({
+                title: formatMessage(
+                    {defaultMessage: '{errorMessage}'},
+                    {errorMessage: API_ERROR_MESSAGE}
+                ),
+                status: 'error'
+            })
+        }
+    }
+
     return (
         <ProductScroller
             ref={ref}
             title={title || recommendations?.displayMessage}
             products={recommendations.recs}
-            onProductClick={(product) =>
-                sendClickReco(
-                    {
-                        recommenderName: recommendations.recommenderName,
-                        __recoUUID: recommendations.recoUUID
-                    },
-                    product
-                )
-            }
             isLoading={loading}
+            productTileProps={(product) => ({
+                onClick: () => {
+                    sendClickReco(
+                        {
+                            recommenderName: recommendations.recommenderName,
+                            __recoUUID: recommendations.recoUUID
+                        },
+                        product
+                    )
+                },
+                enableFavourite: wishlist.isInitialized,
+                isFavourite: !!wishlist.findItemByProductId(product?.productId),
+                onFavouriteToggle: (isFavourite) => {
+                    const action = isFavourite ? addItemToWishlist : removeItemFromWishlist
+                    return action(product)
+                }
+            })}
             {...props}
         />
     )